Clean up addValidator, modifyValidator
Renamed and restructured some things in order to make the functions more
obvious, and to emphasize their structural similarity.

The fact that the `modifier` needs a `bind(this)` (valchemy.js:18) when
the `validator` in `addValidator` does not seems weird to me; I suspect
that they'll converge one way or the other eventually.