WIP: A riddle!
Step 1: Take a look at spec/modifiers/forAttributeSpec.

Un-comment the it-statements one at a time, and try to get them passing.

forAttributeSpec includes a preface at the top to give some context on
the intent behind the feature.

Step 2: For double bonus points, if there's time, uncomment the
commented tests in BasicValidationSpec and try to get those passing too!