Use actual Result objects in attributeErrors
Previously, the values of the `attributeErrors` object on a result were
lists of strings, which were the extracted error messages of validation
results for those attributes. This commit preserves the entire result in
`attributeErrors` instead of just the message strings, which makes it
possible to validate nested objects.

(To see why this was an issue, try checking out to before this commit
and adding the test this commit adds to usageExampleSpec.)

Some related interesting changes:

- When there are no `attributeErrors` on a result object, the value of
  `attributeErrors` is now an empty hash instead of undefined. Having
  this consistency makes it easier to process result objects without
  surprises.
- Extracted a `combineResults` method whose job is just to take two
  result objects, and return one result object which is the combination
  of the two. This drastically simplifies the logic in
  `BasicValidation.validate` for aggregating results, and is hopefully
  easier to reason about than what we were doing before.

Extracting `combineResults` made it apparent that the way attribute
errors are structured is weird. The `forAttribute` modifier flags a
result as being an attribute error by stashing a special `forAttribute`
key on it; this means that the shape of result objects is not
consistent. Results with `forAttribute` keys must be handled one way,
while those without must be handled a different way.

In `combineResults` this problem is solved with the
`normalizeAttributeError` helper function, but I think a better solution
would be to change how the `forAttribute` modifier creates results in
order to make sure that all result objects have the same structure. That
was not done in this commit.