Remove forAttribute key on Result objects
The way we were using the forAttribute key to signal whether a result
was a base or attribute error meant that result objects had inconsistent
structure, which resulted in complexity when we needed to process them.

This commit removes the forAttribute key. Instead, results of validators
that have been run through the forAttribute modifier have the same
structure as other results, they just have an empty `errors` list and an
`attributeErrors` object with a single key-value entry whose key is the
attribute the validator was run on, and whose value is the original
result from the validator. (Look at the diff in forAttributeSpec for
examples of what this looks like concretely.)

Because attribute results now have the same structure as other results,
there is no need for the normalizeAttributeError function in
combineResults.

Also in this commit, I found a much nicer way to combine attributeErrors
objects in combineResults by using the 'customizer' parameter on
lodash's merge function. This cleaned up the module and also made it
more obvious that combineResults is recursive.